Remove unnecessary RPC profiling code after future merge
Now that the futures are consolidated after
https://github.com/pytorch/pytorch/pull/35154, there is no
`torch.distributed.rpc.Future` and we do not need a special path. All futures
can now be profiled through the use of the jit operator defined in
record_function_ops.cpp

As a result, we also get rid of the record_function_ops.h file.
RPC profiling tests are currently disabled, although I re-enabled them locally
to ensure that they still work with this change.

Differential Revision: [D21506091](https://our.internmc.facebook.com/intern/diff/D21506091/)

ghstack-source-id: 103859136
Pull Request resolved: https://github.com/pytorch/pytorch/pull/38255